Issue 4: Documentation on Operations
The block that creates an Operation can set the documentation for
the operation via a new desc method, e.g.:

```ruby
operation :create do
  desc 'Creates a user'
  url '/:username'
  ..
end
```

The existing desc class method added by Seahorse::Model continues to
work, but the new method on Operation takes precedence.

The example app is updated to show both approaches, including using
both in the same model.

I think the class method should be removed from the Model module
but did not want to make this a breaking change.